Update index.rst
Added bullet points for the pages we need to create to form a walk-through experience for new users. Kept the existing content for use as examples, which can be expanded on more over time. Note - this version is unlikely to build in RTD, but I'll fix that in other commits